Feat bugfix/theme-and-ui: Refactor scheme & color changer feature to viewmodel & component, modularize nav drawer
