+++ app/src/main/java/com/digiventure/ventnote/ui/ColorSchemeChoice.kt
@@ -5,82 +5,82 @@ import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import com.digiventure.ventnote.commons.ColorPalletName
 import com.digiventure.ventnote.commons.ColorSchemeName
-import com.digiventure.ventnote.ui.theme.CadmiumGreenDarkPrimary
-import com.digiventure.ventnote.ui.theme.CadmiumGreenDarkPrimaryContainer
-import com.digiventure.ventnote.ui.theme.CadmiumGreenDarkSecondary
-import com.digiventure.ventnote.ui.theme.CadmiumGreenDarkSecondaryContainer
-import com.digiventure.ventnote.ui.theme.CadmiumGreenDarkTertiary
-import com.digiventure.ventnote.ui.theme.CadmiumGreenDarkTertiaryContainer
-import com.digiventure.ventnote.ui.theme.CadmiumGreenLightPrimary
-import com.digiventure.ventnote.ui.theme.CadmiumGreenLightPrimaryContainer
-import com.digiventure.ventnote.ui.theme.CadmiumGreenLightSecondary
-import com.digiventure.ventnote.ui.theme.CadmiumGreenLightSecondaryContainer
-import com.digiventure.ventnote.ui.theme.CadmiumGreenLightTertiary
-import com.digiventure.ventnote.ui.theme.CadmiumGreenLightTertiaryContainer
-import com.digiventure.ventnote.ui.theme.CobaltBlueDarkPrimary
-import com.digiventure.ventnote.ui.theme.CobaltBlueDarkPrimaryContainer
-import com.digiventure.ventnote.ui.theme.CobaltBlueDarkSecondary
-import com.digiventure.ventnote.ui.theme.CobaltBlueDarkSecondaryContainer
-import com.digiventure.ventnote.ui.theme.CobaltBlueDarkTertiary
-import com.digiventure.ventnote.ui.theme.CobaltBlueDarkTertiaryContainer
-import com.digiventure.ventnote.ui.theme.CobaltBlueLightPrimary
-import com.digiventure.ventnote.ui.theme.CobaltBlueLightPrimaryContainer
-import com.digiventure.ventnote.ui.theme.CobaltBlueLightSecondary
-import com.digiventure.ventnote.ui.theme.CobaltBlueLightSecondaryContainer
-import com.digiventure.ventnote.ui.theme.CobaltBlueLightTertiary
-import com.digiventure.ventnote.ui.theme.CobaltBlueLightTertiaryContainer
-import com.digiventure.ventnote.ui.theme.CrimsonDarkPrimary
-import com.digiventure.ventnote.ui.theme.CrimsonDarkPrimaryContainer
-import com.digiventure.ventnote.ui.theme.CrimsonDarkSecondary
-import com.digiventure.ventnote.ui.theme.CrimsonDarkSecondaryContainer
-import com.digiventure.ventnote.ui.theme.CrimsonDarkTertiary
-import com.digiventure.ventnote.ui.theme.CrimsonDarkTertiaryContainer
-import com.digiventure.ventnote.ui.theme.CrimsonLightPrimary
-import com.digiventure.ventnote.ui.theme.CrimsonLightPrimaryContainer
-import com.digiventure.ventnote.ui.theme.CrimsonLightSecondary
-import com.digiventure.ventnote.ui.theme.CrimsonLightSecondaryContainer
-import com.digiventure.ventnote.ui.theme.CrimsonLightTertiary
-import com.digiventure.ventnote.ui.theme.CrimsonLightTertiaryContainer
-import com.digiventure.ventnote.ui.theme.DarkBackground
-import com.digiventure.ventnote.ui.theme.DarkOnBackground
-import com.digiventure.ventnote.ui.theme.DarkOnPrimary
-import com.digiventure.ventnote.ui.theme.DarkOnPrimaryContainer
-import com.digiventure.ventnote.ui.theme.DarkOnSecondary
-import com.digiventure.ventnote.ui.theme.DarkOnSecondaryContainer
-import com.digiventure.ventnote.ui.theme.DarkOnSurface
-import com.digiventure.ventnote.ui.theme.DarkOnSurfaceVariant
-import com.digiventure.ventnote.ui.theme.DarkOnTertiary
-import com.digiventure.ventnote.ui.theme.DarkOnTertiaryContainer
-import com.digiventure.ventnote.ui.theme.DarkOutline
-import com.digiventure.ventnote.ui.theme.DarkOutlineVariant
-import com.digiventure.ventnote.ui.theme.DarkSurface
-import com.digiventure.ventnote.ui.theme.DarkSurfaceVariant
-import com.digiventure.ventnote.ui.theme.LightBackground
-import com.digiventure.ventnote.ui.theme.LightOnBackground
-import com.digiventure.ventnote.ui.theme.LightOnPrimary
-import com.digiventure.ventnote.ui.theme.LightOnPrimaryContainer
-import com.digiventure.ventnote.ui.theme.LightOnSecondary
-import com.digiventure.ventnote.ui.theme.LightOnSecondaryContainer
-import com.digiventure.ventnote.ui.theme.LightOnSurface
-import com.digiventure.ventnote.ui.theme.LightOnSurfaceVariant
-import com.digiventure.ventnote.ui.theme.LightOnTertiary
-import com.digiventure.ventnote.ui.theme.LightOnTertiaryContainer
-import com.digiventure.ventnote.ui.theme.LightOutline
-import com.digiventure.ventnote.ui.theme.LightOutlineVariant
-import com.digiventure.ventnote.ui.theme.LightSurface
-import com.digiventure.ventnote.ui.theme.LightSurfaceVariant
-import com.digiventure.ventnote.ui.theme.PurpleDarkPrimary
-import com.digiventure.ventnote.ui.theme.PurpleDarkPrimaryContainer
-import com.digiventure.ventnote.ui.theme.PurpleDarkSecondary
-import com.digiventure.ventnote.ui.theme.PurpleDarkSecondaryContainer
-import com.digiventure.ventnote.ui.theme.PurpleDarkTertiary
-import com.digiventure.ventnote.ui.theme.PurpleDarkTertiaryContainer
-import com.digiventure.ventnote.ui.theme.PurpleLightPrimary
-import com.digiventure.ventnote.ui.theme.PurpleLightPrimaryContainer
-import com.digiventure.ventnote.ui.theme.PurpleLightSecondary
-import com.digiventure.ventnote.ui.theme.PurpleLightSecondaryContainer
-import com.digiventure.ventnote.ui.theme.PurpleLightTertiary
-import com.digiventure.ventnote.ui.theme.PurpleLightTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenDarkPrimary
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenDarkPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenDarkSecondary
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenDarkSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenDarkTertiary
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenDarkTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenLightPrimary
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenLightPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenLightSecondary
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenLightSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenLightTertiary
+import com.digiventure.ventnote.ui.theme.components.CadmiumGreenLightTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueDarkPrimary
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueDarkPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueDarkSecondary
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueDarkSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueDarkTertiary
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueDarkTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueLightPrimary
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueLightPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueLightSecondary
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueLightSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueLightTertiary
+import com.digiventure.ventnote.ui.theme.components.CobaltBlueLightTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.CrimsonDarkPrimary
+import com.digiventure.ventnote.ui.theme.components.CrimsonDarkPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.CrimsonDarkSecondary
+import com.digiventure.ventnote.ui.theme.components.CrimsonDarkSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.CrimsonDarkTertiary
+import com.digiventure.ventnote.ui.theme.components.CrimsonDarkTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.CrimsonLightPrimary
+import com.digiventure.ventnote.ui.theme.components.CrimsonLightPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.CrimsonLightSecondary
+import com.digiventure.ventnote.ui.theme.components.CrimsonLightSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.CrimsonLightTertiary
+import com.digiventure.ventnote.ui.theme.components.CrimsonLightTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.DarkBackground
+import com.digiventure.ventnote.ui.theme.components.DarkOnBackground
+import com.digiventure.ventnote.ui.theme.components.DarkOnPrimary
+import com.digiventure.ventnote.ui.theme.components.DarkOnPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.DarkOnSecondary
+import com.digiventure.ventnote.ui.theme.components.DarkOnSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.DarkOnSurface
+import com.digiventure.ventnote.ui.theme.components.DarkOnSurfaceVariant
+import com.digiventure.ventnote.ui.theme.components.DarkOnTertiary
+import com.digiventure.ventnote.ui.theme.components.DarkOnTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.DarkOutline
+import com.digiventure.ventnote.ui.theme.components.DarkOutlineVariant
+import com.digiventure.ventnote.ui.theme.components.DarkSurface
+import com.digiventure.ventnote.ui.theme.components.DarkSurfaceVariant
+import com.digiventure.ventnote.ui.theme.components.LightBackground
+import com.digiventure.ventnote.ui.theme.components.LightOnBackground
+import com.digiventure.ventnote.ui.theme.components.LightOnPrimary
+import com.digiventure.ventnote.ui.theme.components.LightOnPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.LightOnSecondary
+import com.digiventure.ventnote.ui.theme.components.LightOnSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.LightOnSurface
+import com.digiventure.ventnote.ui.theme.components.LightOnSurfaceVariant
+import com.digiventure.ventnote.ui.theme.components.LightOnTertiary
+import com.digiventure.ventnote.ui.theme.components.LightOnTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.LightOutline
+import com.digiventure.ventnote.ui.theme.components.LightOutlineVariant
+import com.digiventure.ventnote.ui.theme.components.LightSurface
+import com.digiventure.ventnote.ui.theme.components.LightSurfaceVariant
+import com.digiventure.ventnote.ui.theme.components.PurpleDarkPrimary
+import com.digiventure.ventnote.ui.theme.components.PurpleDarkPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.PurpleDarkSecondary
+import com.digiventure.ventnote.ui.theme.components.PurpleDarkSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.PurpleDarkTertiary
+import com.digiventure.ventnote.ui.theme.components.PurpleDarkTertiaryContainer
+import com.digiventure.ventnote.ui.theme.components.PurpleLightPrimary
+import com.digiventure.ventnote.ui.theme.components.PurpleLightPrimaryContainer
+import com.digiventure.ventnote.ui.theme.components.PurpleLightSecondary
+import com.digiventure.ventnote.ui.theme.components.PurpleLightSecondaryContainer
+import com.digiventure.ventnote.ui.theme.components.PurpleLightTertiary
+import com.digiventure.ventnote.ui.theme.components.PurpleLightTertiaryContainer
 
 object ColorSchemeChoice {
     fun getColorScheme(colorScheme: String, colorPallet: String): ColorScheme {
