+++ app/src/main/java/com/digiventure/ventnote/ui/theme/viewmodel/ThemeBaseVM.kt
@@ -1,0 +1,26 @@
+package com.digiventure.ventnote.ui.theme.viewmodel
+
+import androidx.compose.material3.ColorScheme
+import androidx.compose.runtime.MutableState
+
+interface ThemeBaseVM {
+    /**
+     *
+     * */
+    val currentColorScheme: MutableState<ColorScheme>
+
+    /**
+     *
+     * */
+    val currentColorSchemeName: MutableState<String>
+
+    /**
+     *
+     * */
+    fun updateColorPallet(colorPallet: String)
+
+    /**
+     *
+     * */
+    fun updateColorScheme(colorSchemeName: String)
+}
